redesigned spotipy interaction with main
